requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
ccxt==4.1.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.23
python-binance==1.0.17
aiohttp==3.9.1
schedule==1.2.0

